Move all examples to new examples/ directory (without re-arrange)
From now on, BehTavan aims to be a library instead of a program.
So, existing examples makes no sense to be a main part of the
project structure.

Also, the tiny types.hpp unit is no longer needed, as it is used by
these examples mainly. Plus, there would be no main.cpp for a lib.